chore: migrate base interaction files
